Huge PR with near-total rewrite: CRFs, HTML output, workflow, alouette (#4)
* feat: add MCID and tag

* fix: increase tolerance (for OCR documents)

* feat: reannotate with mcid, mctag

* feat: peaufiner logique de tableau

* chore: mise a jour

* feat: reannotation

* feat: initial implementation of CRF segmentation

* fix: add negative features

* feat: basic features (0.78 macro avg)

* feat: unigram features (0.72 macro avg)

* feat: sync traits entre pdf et csv

* feat: track with mlflow (for the moment)

* feat: ajouter n-grams

* feat: fix delta for top of page

* fix: missing B-TOC

* feat: add BULLET features

* feat: best crf yet (needs more data)

* feat: integrer CRF dans la chaine alexi

* fix: fix annotations and add PIIA

* feat: tweak features

* feat: add default crf model

* fix: corriger des erreurs de traits

* feat: ajouter elements structurels

* chore: add model

* fix: ah much better!

* feat: rule-based baseline

* feat: add bullet feature

* chore: retrain

* fix: model re-entraine, 97 F1 (contre 95 pour regles)

* feat: formatteur xml

* fix: remove Attendu

* fix: add O

* feat: Enumeration -> Liste

* refactor: meilleur feature couleur et refactoring

* chore: mise a jour des csv

* fix: type checking

* fix: add back Tableau

* fix: add numeric bullet type

* fix: corriger plusieurs annotations

* feat: add more data

* feat: mise a jour du training

* feat: retrain including devset

* feat: encore des données (bizarres)

* chore: retrain

* feat: nouvel analyseur

* fix: mineur

* test: enlever tests inutiles

* feat: mettre formattage dans format et ajouter HTML

* feat!: remove legacy tagger

* fix: retrain without +1:bias -1:bias

* feat: sort of handle images (bboxes are mismatched for tables)

* fix!: remove image handling

* chore: reformat

* refactor: update some field names and add json

* fix: is not an article

* test: ajouter test de formattage html

* feat: add -niter (and retrain with best niter)

* feat: formattage d un seul element

* feat!: remove as will rewrite in python

* feat: extraire tableaux/figures avec converteur

* refactor: crf -> segment

* feat: nouvelle extraction tableaux et figures

* test: test data for figures

* feat: incorporer tableaux et figures dans analyse

* refactor: contenus -> contenu

* feat: mostly support tables/figures again

* feat: enlever transitions impossibles et augmenter L1

* revert: unfortunately this does nothing

* feat: generer tous les elements

* feat: turn up the regularization

* fix: correct many Liste labels

* feat: formattage texte

* fix: omettre en tete et pied

* feat: line breaks in text output

* fix: improve text

* chore: update reglements

* feat: add script to get comparable bylaws

* feat: create training data for title extraction

* feat: annotate titles and (some) numbers

* fix: s/SeqTag/seqtag/

* feat: allow for other keys in tagging

* fix: annotate test-title

* fix: seqtag

* refactor: move sequence tags into main training data

* refactor: move sequence tags into main training data, part 2

* feat: use "segtag" and "seqtag" (these are too similar though)

* feat: train model for title/number extraction

* feat: extract title and include in html

* refactor: s/seqtag,segtag/sequence,segment/

* feat: add `alexi label`

* fix: accept iterables in analyser

* feat: output SERAFIM compatible json mostly

* fix: correct an annotation

* feat: tableaux/figures

* fix: extract tables/figures correctly from structure tree

* fix: correct the insertion of images into document structure

* fix: remove bogus words

* fix: handle bogus figures

* fix: remove unused function

* feat: output in SERAFIM format

* fix: fix crashing on serafim output

* feat: update urbanisme

* feat: new version of convert.hs

* fix: try to deal with annexes and empty titles

* feat: extraction en HTML hierarchise

* feat: extraire numeros partout et creer ancrages

* refactor: refactor extraction delements

* feat: enhanced download

* fix: update files

* feat: ameliorer le HTML de sortie

* feat: begin rewrite of index

* fix: generer des pages HTML completes avec titre

* fix: make index/search minimally work again

* feat: emettre des urls pdfs de search

* feat: try to run this as a workflow

+++ data/update_train.py
@@ -2,42 +2,32 @@
 
 import argparse
 import csv
+import logging
 from pathlib import Path
 
 
-def word_eq(r1, r2):
-    return (
-        r1["text"] == r2["text"]
-        and abs(float(r1["x0"]) - float(r2["x0"])) < 1
-        and abs(float(r1["top"]) - float(r2["top"])) < 1
-    )
+def key(row: dict):
+    return (row["text"], row["page"], row["x0"], row["top"])
 
 
 def update_tags(annodir: Path):
     for p in annodir.iterdir():
         if p.suffix != ".csv":
             continue
+        if not Path(p.name).exists():
+            continue
+        old_tags = {}
         with open(p, "rt") as oldfh:
-            old_rows = list(csv.DictReader(oldfh))
-            old_iter = iter(old_rows)
-            old_row = next(old_iter)
+            for row in csv.DictReader(oldfh):
+                old_tags[key(row)] = row["tag"]
+        out_rows = []
         with open(Path(p.name), "rt") as newfh:
             reader = csv.DictReader(newfh)
             fieldnames = reader.fieldnames
-            new_rows = list(reader)
-        out_rows = []
-        for row in new_rows:
-            # Luckily we only *insert* new rows
-            if old_row and word_eq(row, old_row):
-                # Carry over old tag (keep tags on unseen rows)
-                if row["tag"] == "":
-                    row["tag"] = old_row["tag"]
-                try:
-                    old_row = next(old_iter)
-                except StopIteration:
-                    old_row = None
-            out_rows.append(row)
-        assert len(out_rows) == len(new_rows)
+            for row in reader:
+                if not row["tag"]:
+                    row["tag"] = old_tags.get(key(row), "O")
+                out_rows.append(row)
         with open(p, "wt") as outfh:
             writer = csv.DictWriter(outfh, fieldnames=fieldnames)
             writer.writeheader()
@@ -49,6 +39,7 @@ def main():
     parser = argparse.ArgumentParser(description=__doc__)
     parser.add_argument("annodirs", nargs="+", type=Path)
     args = parser.parse_args()
+    logging.basicConfig()
     for d in args.annodirs:
         update_tags(d)
